requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-dialogflow==2.24.1
requests==2.31.0
gunicorn==21.2.0